+++ server/src/main/java/org/opensearch/repositories/blobstore/BlobStoreRepository.java
@@ -114,6 +114,9 @@ import org.opensearch.index.snapshots.blobstore.SlicedInputStream;
 import org.opensearch.index.snapshots.blobstore.SnapshotFiles;
 import org.opensearch.index.store.Store;
 import org.opensearch.index.store.StoreFileMetadata;
+import org.opensearch.index.store.lockmanager.FileLockInfo;
+import org.opensearch.index.store.lockmanager.RemoteStoreLockManagerFactory;
+import org.opensearch.index.store.lockmanager.RemoteStoreMetadataLockManager;
 import org.opensearch.indices.recovery.RecoverySettings;
 import org.opensearch.indices.recovery.RecoveryState;
 import org.opensearch.repositories.IndexId;
@@ -507,8 +510,8 @@ public abstract class BlobStoreRepository extends AbstractLifecycleComponent imp
         executor.execute(ActionRunnable.supply(listener, () -> {
             final long startTime = threadPool.absoluteTimeInMillis();
             final BlobContainer shardContainer = shardContainer(index, shardNum);
-            final BlobStoreIndexShardSnapshots existingSnapshots;
             final String newGen;
+            final BlobStoreIndexShardSnapshots existingSnapshots;
             final String existingShardGen;
             if (shardGeneration == null) {
                 Tuple<BlobStoreIndexShardSnapshots, Long> tuple = buildBlobStoreIndexShardSnapshots(
@@ -561,6 +564,9 @@ public abstract class BlobStoreRepository extends AbstractLifecycleComponent imp
                         + "]. A snapshot by that name already exists for this shard."
                 );
             }
+            // We don't need to check if there exists a shallow snapshot with the same name as we have the check before starting the clone
+            // operation ensuring that the snapshot name is available by checking the repository data. Also, the new clone snapshot would
+            // have a different UUID and hence a new unique snap-N file will be created.
             final BlobStoreIndexShardSnapshot sourceMeta = loadShardSnapshot(shardContainer, source);
             logger.trace("[{}] [{}] writing shard snapshot file for clone", shardId, target);
             INDEX_SHARD_SNAPSHOT_FORMAT.write(
@@ -576,6 +582,91 @@ public abstract class BlobStoreRepository extends AbstractLifecycleComponent imp
                 compressor
             );
             return newGen;
+        }));
+    }
+
+    @Override
+    public void cloneRemoteStoreIndexShardSnapshot(
+        SnapshotId source,
+        SnapshotId target,
+        RepositoryData repositoryData,
+        RepositoryShardId shardId,
+        @Nullable String shardGeneration,
+        RemoteStoreLockManagerFactory remoteStoreLockManagerFactory,
+        ActionListener<String> listener
+    ) {
+        if (isReadOnly()) {
+            listener.onFailure(new RepositoryException(metadata.name(), "cannot clone shard snapshot on a readonly repository"));
+            return;
+        }
+        final IndexId index = shardId.index();
+        final int shardNum = shardId.shardId();
+        final Executor executor = threadPool.executor(ThreadPool.Names.SNAPSHOT);
+        executor.execute(ActionRunnable.supply(listener, () -> {
+            // We don't need to check if there exists a shallow/full copy snapshot with the same name as we have the check before starting
+            // the clone operation ensuring that the snapshot name is available by checking the repository data. Also, the new clone shallow
+            // snapshot
+            // would have a different UUID and hence a new unique shallow-snap-N file will be created.
+            try {
+                final long startTime = threadPool.relativeTimeInMillis();
+                final BlobContainer shardContainer = shardContainer(index, shardNum);
+                RemoteStoreShardShallowCopySnapshot remStoreBasedShardMetadata = loadShallowCopyShardSnapshot(shardContainer, source);
+                String indexUUID = remStoreBasedShardMetadata.getIndexUUID();
+                String remoteStoreRepository = remStoreBasedShardMetadata.getRemoteStoreRepository();
+                RemoteStoreMetadataLockManager remoteStoreMetadataLockManger = remoteStoreLockManagerFactory.newLockManager(
+                    remoteStoreRepository,
+                    indexUUID,
+                    String.valueOf(shardId.shardId())
+                );
+                remoteStoreMetadataLockManger.cloneLock(
+                    FileLockInfo.getLockInfoBuilder().withAcquirerId(source.getUUID()).build(),
+                    FileLockInfo.getLockInfoBuilder().withAcquirerId(target.getUUID()).build()
+                );
+                REMOTE_STORE_SHARD_SHALLOW_COPY_SNAPSHOT_FORMAT.write(
+                    remStoreBasedShardMetadata.asClone(target.getName(), startTime, threadPool.absoluteTimeInMillis() - startTime),
+                    shardContainer,
+                    target.getUUID(),
+                    compressor
+                );
+                return shardGeneration;
+            } catch (Exception e) {
+                logger.info("Exception caught!");
+                List<String> indices = this.getSnapshotInfo(source).indices();
+                List<IndexId> indexIds = repositoryData.resolveIndices(indices);
+                for (IndexId indexId : indexIds) {
+                    logger.info("IndexId: {}", indexId);
+                    IndexMetadata indexMetadata;
+                    try {
+                        indexMetadata = this.getSnapshotIndexMetaData(repositoryData, source, indexId);
+                    } catch (Exception ex) {
+                        continue;
+                    }
+                    if (indexMetadata.getSettings().getAsBoolean(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, false)) {
+                        int numberOfShards = indexMetadata.getNumberOfShards();
+                        for (int shard = 0; shard < numberOfShards; shard++) {
+                            logger.info("Trying to release lock for {} {}", indexId, shard);
+                            final int finalShardId = shard;
+                            String indexUUID = indexMetadata.getIndexUUID();
+                            String remoteStoreRepoForIndex = indexMetadata.getSettings().get(IndexMetadata.SETTING_REMOTE_STORE_REPOSITORY);
+                            RemoteStoreMetadataLockManager remoteStoreMetadataLockManager = remoteStoreLockManagerFactory.newLockManager(
+                                remoteStoreRepoForIndex,
+                                indexUUID,
+                                String.valueOf(finalShardId)
+                            );
+                            try {
+                                remoteStoreMetadataLockManager.release(
+                                    FileLockInfo.getLockInfoBuilder().withAcquirerId(target.getUUID()).build()
+                                );
+                            } catch (Exception ex) {
+                                logger.info("IGNORED EXCEPTION");
+                                // ignoring all exceptions while cleaning up lock files. Uncleaned files will be taken care of either during
+                                // delete/cleanup operation.
+                            }
+                        }
+                    }
+                }
+                throw e;
+            }
         }));
     }
 
@@ -1721,7 +1812,7 @@ public abstract class BlobStoreRepository extends AbstractLifecycleComponent imp
         if (cacheRepositoryData && bestEffortConsistency == false) {
             final BytesReference serialized;
             try {
-                serialized = CompressorFactory.COMPRESSOR.compress(updated);
+                serialized = CompressorFactory.defaultCompressor().compress(updated);
                 final int len = serialized.length();
                 if (len > ByteSizeUnit.KB.toBytes(500)) {
                     logger.debug(
@@ -1757,7 +1848,7 @@ public abstract class BlobStoreRepository extends AbstractLifecycleComponent imp
     }
 
     private RepositoryData repositoryDataFromCachedEntry(Tuple<Long, BytesReference> cacheEntry) throws IOException {
-        try (InputStream input = CompressorFactory.COMPRESSOR.threadLocalInputStream(cacheEntry.v2().streamInput())) {
+        try (InputStream input = CompressorFactory.defaultCompressor().threadLocalInputStream(cacheEntry.v2().streamInput())) {
             return RepositoryData.snapshotsFromXContent(
                 XContentType.JSON.xContent().createParser(NamedXContentRegistry.EMPTY, LoggingDeprecationHandler.INSTANCE, input),
                 cacheEntry.v1()